requirements.txt
selenium
chromedriver-binary-auto

beautifulsoup4
ray
click==8.0
sqlalchemy